requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
filelock==3.10.7
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.2.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
